backend/requirements.txt
# FastAPI & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Celery & Redis
celery==5.3.4
redis==5.0.1

# File Storage
minio==7.2.3
Pillow==10.2.0

# AI/ML - Whisper
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2

# Audio processing
pydub==0.25.1
ffmpeg-python==0.2.0

# Email
aiosmtplib==3.0.1
jinja2==3.1.3

# scheduler
APScheduler==3.10.4

# Utils
python-dotenv==1.0.0
httpx==0.26.0
